Init commit
Experiment trying to make use of GPT to create an OpenAPI 3.1
specification from a Yang file.

When an answer, containing JSON, is received it is validated
to make sure it conform to the OpenAPI 3.1 specification,
and if not, The errors will be returned to the GPT asking it
to correct the errors and return an updated spec. This is
repeated until no more errors are found.